Removed unused code. Organized code.
