+++ openbox/artifact/artifact/settings.py
@@ -51,7 +51,7 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    'user_board.utils.MyAuthMiddleware.MyAuthMiddleware',
+    'user_board.online.MyAuthMiddleware.MyAuthMiddleware',
 ]
 
 ROOT_URLCONF = 'artifact.urls'
