+++ openbox/core/online/flow2.py
@@ -3,7 +3,7 @@ from typing import List
 from ConfigSpace import ConfigurationSpace, Configuration
 
 from openbox import Observation
-from openbox.core.online.utils.base_online_advisor import OnlineAdvisor
+from openbox.core.online.base_online_advisor import OnlineAdvisor
 
 
 class FLOW2(OnlineAdvisor):
