+++ test/reproduction/soc/soc_benchmark_function.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-# from openbox.utils.config_space import ConfigurationSpace, UniformFloatHyperparameter, UniformIntegerHyperparameter, Constant
+# from openbox.online.config_space import ConfigurationSpace, UniformFloatHyperparameter, UniformIntegerHyperparameter, Constant
 from ConfigSpace import ConfigurationSpace, UniformFloatHyperparameter, UniformIntegerHyperparameter, \
     Constant, CategoricalHyperparameter, InCondition, EqualsCondition, UnParametrizedHyperparameter, \
     ForbiddenEqualsClause, ForbiddenInClause, ForbiddenAndConjunction
